Reworked to use RPC API
This changes the sink handler to use the Designate RPC API instead of
the Designate client.  The following changes are included:

CLI program now pulls in the Designate config file and RPC libraries to
support being able to speak to Designate via Rabbit.  This means the CLI
program can now only be run with a valid Designate config and the
ability to speak with Rabbit over AMQP.

Moved all helper functions into the main handler class in order to give
them access to the central_api client provided by the base handler.

Changed to use the format based name generation instead of manually
glomming the name together and then tearing it apart to generate the
fallback name.

Take advantage of managed fields to store the floating IP UUID and to
search and destroy only the relevant records when a floating IP is
disassociated

Added support for processing port.delete message.  This is generated
when the instance is deleted without the floating IP being disassociated
first.  To support acting on this message, the port ID from the original
associate notification is stored in the managed_extra field.

Change-Id: I72d9178e71d075408d4030453e43259badb7459e