Fixed name fallback on duplicate
The fallback behavior was intended to use the fallback record name
format if there was a conflict, but it was checking for a duplicate
record, but duplicates only check for exact data matches, which means
you could end up with multiple records under the same recordset, which
isn't what was intended.

This changes the create_record method to check if a recordset has
records already and if so, throws the CirrusRecordExists exception.
When that exception is thrown when trying to create with the default
name format, it will trigger a retry with the fallback name format.

Change-Id: I0104d203890f1346b2e3147fd447cd51095c3719