Switch to floatingip.delete.end
We don't want to do any updates based on <event>.start events, since
there is no guarantee that the requested change will succeed.  Because
of this, we want to trigger deleting A records on floatingip.delete.end,
not floatingip.delete.start.  This also fixes an existing issue where
the floating ip address is not available in the delete message.

Change-Id: I993cd3b731ece5ebeb011ee2c1682643336eacc0